MLDBFB-615: moved MessageLoop and related classes to libio
