+++ rest/testing/test_peer.h
@@ -16,7 +16,7 @@
 #include "mldb/rest/rest_collection_impl.h"
 #include <chrono>
 #include <thread>
-#include "mldb/soa/service/message_loop.h"
+#include "mldb/io/message_loop.h"
 #include <boost/algorithm/string.hpp>
 #include "mldb/jml/utils/string_functions.h"
