+++ soa/service/testing/runner_test.cc
@@ -22,7 +22,7 @@
 #include "mldb/jml/utils/string_functions.h"
 #include "mldb/jml/utils/vector_utils.h"
 #include "mldb/ext/jsoncpp/value.h"
-#include "mldb/soa/service/message_loop.h"
+#include "mldb/io/message_loop.h"
 #include "mldb/soa/service/runner.h"
 #include "mldb/soa/service/sink.h"
 #include "mldb/types/date.h"
